gradle: remove shadowJar plugin
update of shadowJar plugin on commit d5f1e23a import a new behavior
cannot be disabled currently which collides with the classifier of main jar task.

We have a light use on it to compile local all-deps jar and attach to a release,
so remove it easily and write our own.

Also, re-organize some code in build script.